Add widget configuration for airpods max
