+++ AirpodsPro Battery/AirpodsPro Battery/Assets.xcassets/AirpodsCaseDisconnected.imageset/Contents.json
@@ -1,21 +1,56 @@
 {
   "images" : [
     {
+      "filename" : "airpods(3)-1.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
+      "appearances" : [
+        {
+          "appearance" : "luminosity",
+          "value" : "dark"
+        }
+      ],
+      "filename" : "case-airpods.png",
       "idiom" : "universal",
-      "filename" : "noun_Airpod Case_1801977(1).png",
+      "scale" : "1x"
+    },
+    {
+      "filename" : "airpods(3)@2x-1.png",
+      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "appearances" : [
+        {
+          "appearance" : "luminosity",
+          "value" : "dark"
+        }
+      ],
+      "filename" : "case-airpods@2x.png",
+      "idiom" : "universal",
+      "scale" : "2x"
+    },
+    {
+      "filename" : "airpods(3)@3x-1.png",
+      "idiom" : "universal",
+      "scale" : "3x"
+    },
+    {
+      "appearances" : [
+        {
+          "appearance" : "luminosity",
+          "value" : "dark"
+        }
+      ],
+      "filename" : "case-airpods@3x.png",
       "idiom" : "universal",
       "scale" : "3x"
     }
   ],
   "info" : {
-    "version" : 1,
-    "author" : "xcode"
+    "author" : "xcode",
+    "version" : 1
   }
 }
