+++ AirpodsPro Battery/AirpodsPro Battery/Model/DeviceType.swift
@@ -1,0 +1,15 @@
+//
+//  DeviceType.swift
+//  AirpodsPro Battery
+//
+//  Created by Mohamed Arradi on 22/08/2021.
+//  Copyright © 2021 Mohamed Arradi. All rights reserved.
+//
+
+import Foundation
+
+enum DeviceType: Int {
+    case airpods
+    case headset
+    case unknown
+}
