+++ AirpodsPro Battery/AirpodsPro Battery/Views/BatteryAirpodsView.swift
@@ -1,16 +1,16 @@
 //
-//  BatteryView.swift
+//  BatteryAirpodsView.swift
 //  AirpodsPro Battery
 //
-//  Created by Mohamed Arradi on 21/11/2019.
-//  Copyright © 2019 Mohamed Arradi. All rights reserved.
+//  Created by Mohamed Arradi on 22/08/2021.
+//  Copyright © 2021 Mohamed Arradi. All rights reserved.
 //
 
 import Foundation
 import Cocoa
 
 class BatteryAirpodsView: NSView {
-    
+
     @IBOutlet weak var rightEarBatteryLevelLabel: NSTextField!
     @IBOutlet weak var leftEarBatteryLevelLabel: NSTextField!
     @IBOutlet weak var caseBatteryLevelLabel: NSTextField!
@@ -20,36 +20,16 @@ class BatteryAirpodsView: NSView {
     @IBOutlet weak var caseImageView: NSImageView!
     @IBOutlet weak var leftEarImageView: NSImageView!
     @IBOutlet weak var rightEarImageView: NSImageView!
-    @IBOutlet weak var headsetBatteryView: BatteryHeadsetView!
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-    }
-    
-    func updateViewData(_ viewModel: AirPodsBatteryViewModel?) {
-        
-        guard let airpodsBatteryViewModel = viewModel else {
-            return
-        }
-        
-        if airpodsBatteryViewModel.airpodsInfo != nil {
-            updateData(airpodsBatteryViewModel: airpodsBatteryViewModel)
-            headsetBatteryView.isHidden = true
-        } else if airpodsBatteryViewModel.headsetInfo != nil {
-            headsetBatteryView.isHidden = false
-            headsetBatteryView.updateViewData(viewModel)
-        }
-    }
-    
-    fileprivate func updateData(airpodsBatteryViewModel: AirPodsBatteryViewModel) {
-        
+
+    func updateData(airpodsBatteryViewModel: AirPodsBatteryViewModel) {
+
         caseBatteryLevelLabel.stringValue = airpodsBatteryViewModel.airpodsInfo?.caseDisplayBatteryValue ?? "--"
         batteryLevelCaseProgressBar.progressValue = airpodsBatteryViewModel.airpodsInfo?.caseBatteryValue ?? 0.0
         leftEarBatteryLevelLabel.stringValue = airpodsBatteryViewModel.airpodsInfo?.leftDisplayBatteryValue ?? "--"
         batteryLevelLeftProgressBar.progressValue = airpodsBatteryViewModel.airpodsInfo?.leftBatteryValue ?? 0.0
         rightEarBatteryLevelLabel.stringValue = airpodsBatteryViewModel.airpodsInfo?.rightDisplayBatteryValue ?? "--"
         batteryLevelRightProgressBar.progressValue = airpodsBatteryViewModel.airpodsInfo?.rightBatteryValue ?? 0.0
-        
+
         switch airpodsBatteryViewModel.connectionStatus {
         case .connected:
             leftEarImageView.image = NSImage(imageLiteralResourceName: "LeftAirpodEar")
